Update s2_order to save all components of the order
- Save all components of the order to JSON (available, ordered and notordered)
NOTE: this change of behaviour makes so that when using safe_is_online on the expoted json, availability of all the images in the INITIAL list is check , and not only availability of the ordered products. I think this makes more sense, because in this way it is easier to understand when all the initial list has been properly ordered and re-set online. 
- Allow using a json file to re-do an order, specifying if re-ordering already ordered datasets or only order the ones identified as "notordered" (based on new argument "reorder"
- Provide message about HOW to re-place order based on JSON
- update tests to check the new structure and functioning
- Add reporting message to safe_is_online